refactor: extract home sections
